Update gcd.tex
I think it's generally good, besides minor (wording/phrasing) edits which I made directly, everything else is in comments. Main things are:
a. we should standardise how we refer to it: judgements of lying/deception/veracity/honesty/dishonesty
b. intro could do with some reference to literature on time course of gesture comprehension (can be outside of deception)
c. we definitely need to discuss the rule-of-thumb/speaker-specific modelling hypotheses at the end, given the framing in the intro